Move release templates from internal to pkg.
